객체 미용 체조 2원칙 적용 (switch case 적용하기)
